Update mod_ack to be a gen_server process and the retry logic in the module.
- Made the module to be its own gen_server process so that it can automatically retry sending
messages or send them to offline when the client is no longer online.
- Modified the retry logic so that after we hit the max number of retries due to timeout or
if we hit the limit of max number of items: we direct the packet to offline_msg to retry sending
the packet later when the user comes back online.
- If the user goes offline for some reason when we retry sending the message: the message will automatically
invoke the offline_message_hook to store the message and we will then stop retrying until the user comes back online.
- For each of the hooks in the module, the gen_server callbacks in this module will be called.