Added mod_contacts to handle user contacts
- Previously, we had a module mod_phone_number_normalization to normalize all the contacts.
The client would then individually create pubsub nodes, set affiliations and subscribe to its friends.
- We simplify this whole process now: the client will send a list of contacts to the server, which are stored in a table.
The server handles the process of subscribing the user to its friends contacts, setting affiliations etc.
Server also handles when the client deletes a contact by unsubscribing from relevant nodes and removing the contact from affiliation.
- mod_contacts will now handle iq's with a new namespace halloapp:user:contacts and respond appropriately.
- mod_contacts_mnesia is used to handle all the mnesia related queries with the table of user_contacts.
- mod_contacts will now on receiving the list of contacts, normalizes these phone numbers using the phone_number_util module,
inserts these contacts of the users into an mnesia table, checks if the user and
its contacts are connected on halloapp or not, generates user ids for the user and its contact
if necessary, sets the appropriate affiliations to the users pubsub nodes,
subscribes the user to all the pubsub nodes of its contacts and subscribes all the friend contacts to the user's pubsub node.
- The module also notifies the clients when a friend of the client joins/leaves halloapp.
- exported useful functions in mod_pubsub and mod_contacts.
- Removed mod_phone_number_normalization since its no longer necessary.