Upload server with internal domain name
The internal domain name needs to be resolve to internal IP address
of the upload server. It is important to send the request from
ejabberd to the upload server using the internal ip address in
order for security group based access from ejabberd to tusd running
on the upload server.