Adding User counters. (#243)
* Adding User counters.

Adding 2 counters for accounts and registrations backed by redis.
We have 16K counters (one counter for each Redis Slot).
When account is created we increment the corresponding counter
for the redis slow this user is stored. To do this we precompute
a table to from slot to key that will map on this redis slot.

Currently the counting code is a bit unoptimized because it queries
all 16K counters one at a time. To do better we need changes to
the redis client code.

* Add migration function to fix the counters.

The migration function scans the redis cluster node by node
looking for account and deleted_account keys.

* Switching to fold

maps:fold i guess will not produce a new map

* Deleting the script.

Moving the script to another pull request, as requested.

* Fix small crash and update the deps.

Fixing a bug i introduced when switching from maps:map
to maps:fold.

Also updating the dependency.

* Fix small unit test bug.

Just a small fix.