Second version of the redis_hash.escript
I tried to address some of the comments in the CR
but I don't really want to spend to much time on this code
since it was just a quick script to get me a mapping
and we don't really plan on running it. I just wanted to commit it.