Rename the redis crc16 module to crc16_redis
We had a bug where the new module crc16 shadowed
the crc16 module we were using for model_phone hashing.
This caused the produciton issue where we could not find the old
accounts.

Also added tests to make sure in the future the crc16 modules
continue to produce the same results we expect.

Also a performance test to compare the crc16 (implemented as nif)
vs the crc16_redis (erlang).