Add register_user hook and modify notify_contact_about_user() to only include AvatarId when role is friends (#275)
There was a bug that the server will not notify client A when client B joins HalloApp in the case that A has B's contact but B does not have A's contact. To solve this, add 'register_user' hook and modify 'notify_contact_about_user'. In 'notify_contact_about_user', register process won't send avatarid since A and B are not friends yet.