Add register_user hook and modify notify_contact_about_user() to only include AvatarId when role is friends (#275)
There was a bug that the server will not notify client A when client B joins HalloApp in the case that A has B's contact but B does not have A's contact. To solve this, add 'register_user' hook and modify 'notify_contact_about_user'. In 'notify_contact_about_user', register process won't send avatarid since A and B are not friends yet.

Admin API to trace uid.

Broadcast the update to all nodes in the cluster.
Check every 10 minutes to make we are in sync with the
redis data.

Add ejabberd_admin phone trace commands.

Implement the phone tracing functionality.
When phone is traced we find the uid to trace and add it
to the list of uids. We also save the phones to be traced in redis.

Future work is to listen on register_user and remove_user hooks.