Migrate client_version from Mnesia to Redis Part1.
Here we define the client_version model for redis (+ tests)
We start writing data to the redis model, but rely on the mnesia data
We have migration function to call that will copy old versions to redis.

After this gets deployed and we do the migration we should not be
getting any errors.

Stage 2 will be to switch the reads and writes to redis first,
Stage 3 stop writing to mnesia, cleanup the code

Fix old name.

Fix old name.