Try to make the priv dir path work in all envs. (#791)
* Try to make the priv dir path work in all envs.

We had some custom code that fixed the priv path when
running unit tests. Unit tests run in .eunit folder.
So to get to the lua scripts or other files in priv folder
we were adding .. in 3 different places.
The problem is when we run the ct tests the priv_dir should
be just where it is without the ..

This change attempts to fix this.