Make sure Metric argument in stat:count is string.
The Metric argument is in the count functions was supposed to be string
In some places we were passing atoms. I fixed the callers but
also added warning in the count API and converted to string.

Also removed count_d and just did count function. There is no longer
option to omit the count and specify tags.

Add warning to see what names are not working