Switch to redis in mod_client_version.
Start using the redis data only
Make the 30 days be constant not config option
Try to debug why we get request with "Halloapp/Android0.93"
"Android0.93" and "0.93" all at the same time.
The closing connection logic does not really make sense.
It also does not make sense for this code to try to close the connection
given the c2s process do it also..

* Remove connection closing if the version is too old.
We will close the connection just when they make a new one.