Model Sessions
New redis model for storing session information.
Im storing 2 structure one is simple uid to pid
which should be enough for our needs. One we know the PID
we can always ask for whatever information we want about this sessions
using an rpc.
Second structure is storing number of sessions like the current mnesia.
The plan is to migrate and compare and slowly try to get rid of the
multi sessions code.