Allow whisper keys to be passed during regist.
This change allows whisper keys to be passed
as part of the registration API.
The spec was already aggreed in the server repository.
New parameters are now optional but will become mandatory
once clients update.

Adding more checks and errors related to the whisper keys.
Also optimizing the error handling

notify other of keys update