Cleanup stale sessions from redis on read.
Mnesia session table had code that will read
all the sessions on startup and try to clean up old
sessions for this node. This code would not be able to scale.
Instead with the sessions in redis we tolerate having
stale sessions in the redis. We will be checking if the
PIDs are still alive and cleaning up the old ones.

Using rpc to check if the process is alive on the other nodes.
Set rpc timeout of 5s.
Default timeout for rpc is inf