Switch user traces to new format.
Switching the user traces to pb structures and base64 encoded protobufs
binaries
Also moving the hooks that intercept those packets right at the moment
of encoding and decoding so we get the real bytes we send on the
network.
Passing both the encoded and unencoded forms of the packet to mod_trace

TODO: want to add test for the new logic.