Fix up the shutdown logic.
We want to make sure the c2s process terminate
before the ejabberd_sm does and we are able to cleanup
the session and ets tables.

In this change we first ask ejabberd_sm to terminate all the c2s
processes. We give some grace period for all the c2s process to finish
terminating and then we do one final redis session cleanup.

Move ejabberd_sm:stop lower in the shutdown order