Make sure the redis start early on.
In our old code we starting redis when, mod_redis was started.
This is after the ejabberd_sm starts. We should start our
connections to redis early in the startup process.

Most of the code from mod_redis went into ha_redis.
Bunch of the tests were calling 2 times the redis_sup to start.