Erlang based stress test framework.
* ha_bot is a gen_server that simulates a client.
ha_bot uses the ha_client and registration_client to communicate with
the backend.
* ha_bot is given config that tells the bot what actions to perform and
how often. Currently the bot is only able to do registrations.
* bot actions are defined like this:
   action_registration => {Frequency, Args}

   Frequency is number of operations per second the bot will do. It
   could be floating point number (eg 0.1 means 1 action per 10s on avg)

* N bots are started together and controlled by gen_server process
ha_bot_farm. The bot farm gen_server processes are started and
controlled via the API defined in ha_stress

Example:

    ha_stress:start_link().
    ha_stress:start_bots(10, 2, [node()]). % start 10 bots total split into 2
        % farms, both running on node()
    ha_stress:set_conf(ha_stress:get_conf()). % tell all the bots the config
    ha_stress;set_bots(100). % incrase the number of bots to 100

Just check if the function exists and is exported.

Instead of maintaining list of implemented actions just check if the
function is exported.