Implement pong and less logs.
Implementing a response to server ping requests.
Server will send a pong iq to the client to make sure the client is
still using the connection. We were not implementing a response
so the server was closing the connection during long run idle stress
tests.

Also lowering the log level to debug and commenting out many of the logs
in ha_client. They were useful while we were still developing the
client. Now they are not that useful. Debug level is ok for running the
ct tests, but in the stress test I will run it with info.