migrate invites.
We had older migration of keys in the invite system that we did not
complete. This migration will complete the migration to the new keys.
Once this is complete we will remove the old code in model_invites.

Once this is done, I will write some code about expiring invites in 60
days.