Switch some supervisors to supervisor3 (#1775)
Supervisor3 is copy of the build supervisor with some important changes.
It allows us to restart the children with some delay indefinetely. This
is the behaviour I think we want for things like modules.

We should probably change also redis_sup to supervisor3 also. I flagged
all the supervisors currently in ejabberd with a TODO so we can
eventually review and remove some of them or switch them to supervisor3.