Trying to do a webpage for the invite appclip. (#2258)
Trying to do a webpage for the invite appclip.

I think this the right way to do the appclip. We keep it on halloapp.com
but add the meta banner only on the invite page.
We will need Duygu to design some sort of a web page for the group
invite.